Move instrumentation options into their own .h/.cpp files.
